feat: added authentication to views

+++ contacts/views.py
@@ -15,7 +15,6 @@ from .serializers import (ContactDetailSerializer, ContactNameSerializer,
 class CreateListContactView(ListCreateAPIView):
     """Create a contact or list all contacts"""
     serializer_class = ContactsSerializer
-    queryset = Contacts.objects.all()
     permission_classes = [IsAuthenticated]
 
     def create(self, request, *args, **kwargs):
@@ -25,17 +24,30 @@ class CreateListContactView(ListCreateAPIView):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+    def get_queryset(self):
+        user = self.request.user
+        return Contacts.objects.filter(owner=user)
+
 class ContactDetailView(APIView):
-    permission_classes = [IsAuthenticated]
 
     """Get detailed contact or delete"""
     def get(self, request, contact_id):
         contact = get_object_or_404(Contacts, id=contact_id)
+        if request.user.id != contact.owner.id:
+            return Response(
+                "You don't have access to view this contact.",
+                status=status.HTTP_401_UNAUTHORIZED
+            )
         serializer = ContactDetailSerializer(contact, many=False)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def delete(self, request, contact_id):
         contact = get_object_or_404(Contacts, id=contact_id)
+        if request.user.id != contact.owner.id:
+            return Response(
+                "You don't have access to view this contact.",
+                status=status.HTTP_401_UNAUTHORIZED
+            )
         contact.delete()
         return Response(f"Contact {contact.name} deleted.", status=status.HTTP_200_OK)
 
@@ -45,14 +57,20 @@ class CreatePhoneNumber(CreateAPIView):
 
 class CreateEmail(CreateAPIView):
     serializer_class = EmailSerializer
+    permission_classes = [IsAuthenticated]
 
 class UpdatePhoneNumber(APIView):
     serializer_class = PhoneNumberSerializer
-    permission_classes = [IsAuthenticated]
 
     def put(self, request, phone_number_id):
         number = get_object_or_404(PhoneNumbers, id=phone_number_id)
-        serializer = self.serializer_class(number, data=request.data)
+        if request.user.id != number.contact.owner.id:
+            return Response(
+                "You don't have access to view this contact.",
+                status=status.HTTP_401_UNAUTHORIZED
+            )
+        serializer = self.serializer_class(number, data=request.data,
+                                    context={'phonenumber_id': phone_number_id})
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
@@ -60,6 +78,11 @@ class UpdatePhoneNumber(APIView):
 
     def delete(self, request, phone_number_id):
         number = get_object_or_404(PhoneNumbers, id=phone_number_id)
+        if request.user.id != number.contact.owner.id:
+            return Response(
+                "You don't have access to view this contact.",
+                status=status.HTTP_401_UNAUTHORIZED
+            )
         number.delete()
         return Response(
             f"Phone number {number.phoneNumber} has been deleted.",
@@ -72,7 +95,12 @@ class UpdateEmail(APIView):
 
     def put(self, request, email_id):
         email = get_object_or_404(Emails, id=email_id)
-        serializer = self.serializer_class(email, data=request.data)
+        if request.user.id != email.contact.owner.id:
+            return Response(
+                "You don't have access to view this contact.",
+                status=status.HTTP_401_UNAUTHORIZED
+            )
+        serializer = self.serializer_class(email, data=request.data, context={'email_id': email_id})
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
@@ -80,6 +108,11 @@ class UpdateEmail(APIView):
 
     def delete(self, request, email_id):
         email = get_object_or_404(Emails, id=email_id)
+        if request.user.id != email.contact.owner.id:
+            return Response(
+                "You don't have access to view this contact.",
+                status=status.HTTP_401_UNAUTHORIZED
+            )
         email.delete()
         return Response(
             f"Email {email.email} has been deleted.",
@@ -93,7 +126,8 @@ class UpdateContactName(UpdateAPIView):
     lookup_field = "id"
 
     def get_queryset(self):
-        return Contacts.objects.all()
+        user = self.request.user
+        return Contacts.objects.filter(owner=user)
 
 class UpdateProfilePicutre(UpdateAPIView):
     permission_classes = [IsAuthenticated]
@@ -102,4 +136,5 @@ class UpdateProfilePicutre(UpdateAPIView):
     lookup_field = "id"
 
     def get_queryset(self):
-        return Contacts.objects.all()
+        user = self.request.user
+        return Contacts.objects.filter(owner=user)
